long long

+++ user/jsvar.c
@@ -634,7 +634,7 @@ JsVar *jsvNewFromFloat(JsVarFloat value) {
   var->varData.floating = value;
   return var;
 }
-JsVar *jsvNewFromLongInteger(int64_t value) {
+JsVar *jsvNewFromLongInteger(long long value) {
 //  if (value>=-2147483648LL && value <=2147483647LL)
   if (value >= INT32_MIN && value <= INT32_MAX)
     return jsvNewFromInteger((JsVarInt)value);
@@ -1369,7 +1369,7 @@ JsVarInt jsvGetInteger(const JsVar *v) {
       return jsvGetIntegerAndUnLock(jsvSkipNameAndUnLock(jsvGetArrayItem(v,0)));
     if (jsvIsFloat(v)) {
       if (isfinite(v->varData.floating))
-        return (JsVarInt)(int64_t)v->varData.floating;
+        return (JsVarInt)(long long)v->varData.floating;
       return 0;
     }
     if (jsvIsString(v) && jsvIsStringNumericInt(v, true/* allow decimal point*/)) {
@@ -1380,13 +1380,13 @@ JsVarInt jsvGetInteger(const JsVar *v) {
     return 0;
 }
 
-int64_t jsvGetLongInteger(const JsVar *v) {
-  if (jsvIsInt(v)) return (int64_t)jsvGetInteger(v);
-  return (int64_t)jsvGetFloat(v);
+long long jsvGetLongInteger(const JsVar *v) {
+  if (jsvIsInt(v)) return (long long)jsvGetInteger(v);
+  return (long long)jsvGetFloat(v);
 }
 
-int64_t jsvGetLongIntegerAndUnLock(JsVar *v) {
-  int64_t i = jsvGetLongInteger(v);
+long long jsvGetLongIntegerAndUnLock(JsVar *v) {
+  long long i = jsvGetLongInteger(v);
   jsvUnLock(v);
   return i;
 }
@@ -2516,9 +2516,9 @@ JsVar *jsvMathsOp(JsVar *a, JsVar *b, int op) {
             JsVarInt da = jsvGetInteger(a);
             JsVarInt db = jsvGetInteger(b);
             switch (op) {
-                case '+': return jsvNewFromLongInteger((int64_t)da + (int64_t)db);
-                case '-': return jsvNewFromLongInteger((int64_t)da - (int64_t)db);
-                case '*': return jsvNewFromLongInteger((int64_t)da * (int64_t)db);
+                case '+': return jsvNewFromLongInteger((long long)da + (long long)db);
+                case '-': return jsvNewFromLongInteger((long long)da - (long long)db);
+                case '*': return jsvNewFromLongInteger((long long)da * (long long)db);
                 case '/': return jsvNewFromFloat((JsVarFloat)da/(JsVarFloat)db);
                 case '&': return jsvNewFromInteger(da&db);
                 case '|': return jsvNewFromInteger(da|db);
